requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# AI/ML
torch>=2.0.0
transformers>=4.40.0
Pillow>=10.0.0

# L0 Object Detection (YOLO)
ultralytics>=8.0.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
